requirements.txt
streamlit==1.35.0
streamlit-option-menu==0.3.6
pandas==2.2.2
Pillow==10.3.0
PyPDF2==3.0.1
reportlab==4.1.0
plotly==5.22.0
XlsxWriter==3.2.0

